Many changes towards using the new Neo4j 3.0 Bolt interface for all database operations. Work in progress, do not use!

+++ structr-core/src/main/java/org/structr/core/entity/ResourceAccess.java
@@ -156,7 +156,7 @@ public class ResourceAccess extends AbstractNode {
 		}
 
 		if (cachedPosition != null) {
-			return cachedPosition;
+			return cachedPosition.intValue();
 		}
 
 		return 0;
